changed to floats

+++ RaccoonEngine/src/Camera.java
@@ -8,23 +8,23 @@ public class Camera implements KeyListener {
 	public static boolean left_once, right_once, forward_once, back_once, enter_once, space_once, ctrl_once, strafeleft_once, straferight_once, first_once, second_once, third_once, pgup_once, pgdn_once, fourth_once, f4_once;
 	private boolean left_once_flag, right_once_flag, forward_once_flag, back_once_flag, enter_once_flag, space_once_flag, ctrl_once_flag, strafeleft_once_flag, straferight_once_flag, first_once_flag, second_once_flag, third_once_flag, pgup_once_flag, pgdn_once_flag, fourth_once_flag, f4_once_flag;
 	private boolean fullscreen = false;
-	public static double direction_rad;
-	public static double TURN_SPEED = 0.03;
-	public static double MOVE_SPEED = 0.2;
-	public static double JUMP_UP_SPEED = 0.2;
-	public static double CROUCHING_SPEED = 0.5;
-	public static double gravity_up_multiplier = 1;
-	public static double gravity_down_multiplier = 1;
-	public static double player_x, player_y, player_z;
+	public static float direction_rad;
+	public static float TURN_SPEED = 0.03f;
+	public static float MOVE_SPEED = 0.2f;
+	public static float JUMP_UP_SPEED = 0.2f;
+	public static float CROUCHING_SPEED = 0.5f;
+	public static float gravity_up_multiplier = 1;
+	public static float gravity_down_multiplier = 1;
+	public static float player_x, player_y, player_z;
 	public static int player_sector;
-	public static double retina_dist = 8;
-	public static double player_height = 2;
+	public static float retina_dist = 8;
+	public static float player_height = 2;
 	private boolean jumping_in_progress = false;
-	private double crouch_diff_height = player_height/2;
-	private double jump_diff_height = player_height/4;
+	private float crouch_diff_height = player_height/2;
+	private float jump_diff_height = player_height/4;
 	private boolean jumping_down_flag = false;
 	private boolean crouching_in_progress = false;
-	private double buffer_dist = 0.2;
+	private float buffer_dist = 0.2f;
 	
 
 	public Camera() {
@@ -227,9 +227,9 @@ public class Camera implements KeyListener {
 		}
 	}
 
-	private boolean is_collision(Sector cur_sector, double val_to_add_x, double val_to_add_z) {
-		double temp_x = player_x;
-		double temp_z = player_z;
+	private boolean is_collision(Sector cur_sector, float val_to_add_x, float val_to_add_z) {
+		float temp_x = player_x;
+		float temp_z = player_z;
 		
 		temp_x += val_to_add_x;
 		temp_z += val_to_add_z;
@@ -270,52 +270,52 @@ public class Camera implements KeyListener {
 		Sector cur_sector = Screen.sectorMap.get(player_sector);
 
 		if (right) {
-			direction_rad = (direction_rad - TURN_SPEED) % (2*Math.PI);
+			direction_rad = (float) ((direction_rad - TURN_SPEED) % (2*Math.PI));
 			if (direction_rad < 0) {
 				direction_rad += 2 * Math.PI;
 			}
 		}
 		else if (left) {
-			direction_rad = (direction_rad + TURN_SPEED) % (2*Math.PI);
+			direction_rad = (float) ((direction_rad + TURN_SPEED) % (2*Math.PI));
 		}
 		if (forward) {			
-			if (!is_collision(cur_sector, Math.cos(direction_rad) * (MOVE_SPEED + buffer_dist), 0)) {
+			if (!is_collision(cur_sector, (float)Math.cos(direction_rad) * (MOVE_SPEED + buffer_dist), 0)) {
 				player_x += Math.cos(direction_rad) * MOVE_SPEED;
 			}
-			if (!is_collision(cur_sector, 0, Math.sin(direction_rad) * (MOVE_SPEED + buffer_dist))) {
+			if (!is_collision(cur_sector, 0, (float)Math.sin(direction_rad) * (MOVE_SPEED + buffer_dist))) {
 				player_z += Math.sin(direction_rad) * MOVE_SPEED;
 			}
 		}
 		else if (back) {
-			if (!is_collision(cur_sector, -Math.cos(direction_rad) * (MOVE_SPEED + buffer_dist), 0)) {
+			if (!is_collision(cur_sector, (float) -Math.cos(direction_rad) * (MOVE_SPEED + buffer_dist), 0)) {
 				player_x -= Math.cos(direction_rad) * MOVE_SPEED;
 			}
-			if (!is_collision(cur_sector, 0, -Math.sin(direction_rad) * (MOVE_SPEED + buffer_dist))) {
+			if (!is_collision(cur_sector, 0, (float) -Math.sin(direction_rad) * (MOVE_SPEED + buffer_dist))) {
 				player_z -= Math.sin(direction_rad) * MOVE_SPEED;
 			}
 		}
 		if (straferight) {
-			if (!is_collision(cur_sector, Math.cos(direction_rad - Math.PI/2) * (MOVE_SPEED + buffer_dist), 0)) {
+			if (!is_collision(cur_sector, (float)Math.cos(direction_rad - Math.PI/2) * (MOVE_SPEED + buffer_dist), 0)) {
 				player_x += Math.cos(direction_rad - Math.PI/2) * MOVE_SPEED;
 			}
-			if (!is_collision(cur_sector, 0, Math.sin(direction_rad - Math.PI/2) * (MOVE_SPEED + buffer_dist))) {
+			if (!is_collision(cur_sector, 0, (float)Math.sin(direction_rad - Math.PI/2) * (MOVE_SPEED + buffer_dist))) {
 				player_z += Math.sin(direction_rad - Math.PI/2) * MOVE_SPEED;
 			}
 		}
 		else if (strafeleft) {
-			if (!is_collision(cur_sector, Math.cos(direction_rad + Math.PI/2) * (MOVE_SPEED + buffer_dist), 0)) {
+			if (!is_collision(cur_sector, (float)Math.cos(direction_rad + Math.PI/2) * (MOVE_SPEED + buffer_dist), 0)) {
 				player_x += Math.cos(direction_rad + Math.PI/2) * MOVE_SPEED;
 			}
-			if (!is_collision(cur_sector, 0, Math.sin(direction_rad + Math.PI/2) * (MOVE_SPEED + buffer_dist))) {
+			if (!is_collision(cur_sector, 0, (float)Math.sin(direction_rad + Math.PI/2) * (MOVE_SPEED + buffer_dist))) {
 				player_z += Math.sin(direction_rad + Math.PI/2) * MOVE_SPEED;
 			}
 		}
 
-		double sector_h = cur_sector.floor_height;
-		double player_h_limit = sector_h+player_height;
-		double player_f_limit = sector_h+crouch_diff_height;
-		double move_up_speed = JUMP_UP_SPEED*gravity_up_multiplier;
-		double move_dn_speed = JUMP_UP_SPEED*gravity_down_multiplier;
+		float sector_h = cur_sector.floor_height;
+		float player_h_limit = sector_h+player_height;
+		float player_f_limit = sector_h+crouch_diff_height;
+		float move_up_speed = JUMP_UP_SPEED*gravity_up_multiplier;
+		float move_dn_speed = JUMP_UP_SPEED*gravity_down_multiplier;
 
 		if (pgup_once) {
 			jumping_in_progress = true;
