+++ RaccoonEngine/src/Portal.java
@@ -1,16 +1,16 @@
 
 public class Portal {
-	double x1, z1, x2, z2;
+	float x1, z1, x2, z2;
     int sectorA;
     int sectorB;
     String portalBottomTexture;
-    double portalBottomBrightness;
+    float portalBottomBrightness;
     String portalMiddleTexture;
-    double portalMiddleBrightness;
+    float portalMiddleBrightness;
     String portalTopTexture;
-    double portalTopBrightness;
+    float portalTopBrightness;
     
-    public Portal(double x1, double z1, double x2, double z2, int sectorA, int sectorB, String portalBottomTexture, double portalBottomBrightness, String portalMiddleTexture, double portalMiddleBrightness, String portalTopTexture, double portalTopBrightness) {
+    public Portal(float x1, float z1, float x2, float z2, int sectorA, int sectorB, String portalBottomTexture, float portalBottomBrightness, String portalMiddleTexture, float portalMiddleBrightness, String portalTopTexture, float portalTopBrightness) {
     	this.x1 = x1;
     	this.z1 = z1;
     	this.x2 = x2;
